feature/PDI-169: Promote bank-frick v3/Deprecate bank-frick v2 (#2412)
* Fix a typo in a config error message

* Drop bank-frick (v2)

* Rename bank-frick-test to bank-frick, update dependencies

* Add changeset

* Remove duplicate bank-frick entries from tsconfig
