+++ .changeset/sweet-swans-repeat.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/bank-frick-adapter': major
+---
+
+Converted to a v3 adapter
