+++ packages/core/legos/tsconfig.json
@@ -44,9 +44,6 @@
       "path": "../../sources/bank-frick"
     },
     {
-      "path": "../../sources/bank-frick-test"
-    },
-    {
       "path": "../../sources/bea"
     },
     {
