+++ packages/scripts/src/generate-readme/readmeBlacklist.json
@@ -70,7 +70,6 @@
     "the-graph",
     "token-allocation",
     "vesper",
-    "xsushi-price",
-    "bank-frick-test"
+    "xsushi-price"
   ]
 }
