--- packages/sources/bank-frick-test/CHANGELOG.md
@@ -1,41 +1,0 @@
-# @chainlink/bank-frick-test-adapter
-
-## 1.0.4
-
-### Patch Changes
-
-- a8a33addb: accounts and authorize now use the same URL (\$API_ENDPOINT)
-
-## 1.0.3
-
-### Patch Changes
-
-- b98296089: Bumped v3 framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- 43062ec95: Update to latest framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- Updated dependencies [fc46b78fc]
-- Updated dependencies [1de0689c6]
-  - @chainlink/ea-bootstrap@2.25.0
-  - @chainlink/ea-test-helpers@1.4.1
-
-## 1.0.0
-
-### Major Changes
-
-- b1f5a7113: Initial version of adapter
-
-### Patch Changes
-
-- Updated dependencies [e8576df4e]
-- Updated dependencies [842651f93]
-  - @chainlink/ea-bootstrap@2.24.0
-  - @chainlink/ea-test-helpers@1.4.1
