--- packages/sources/bank-frick-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { accountsRestEndpoint as accounts } from './accounts'
