--- packages/sources/bank-frick-test/test-payload.json
@@ -1,5 +1,0 @@
-{
- "requests": [{
-  "ibanIDs": ["LI6808811000000045345"]
- }]
-}
