--- packages/sources/bank-frick/schemas/env.json
@@ -1,32 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/bank-frick-adapter.json",
-  "title": "Bank Frick Adapter",
-  "description": "Adapter for fetching account information from Bank Frick",
-  "required": ["API_KEY", "PRIVATE_KEY"],
-  "type": "object",
-  "properties": {
-    "API_ENDPOINT": {
-      "type": "string",
-      "description": "The URL for the Bank Frick API.",
-      "default": "https://olbsandbox.bankfrick.li/webapi/v2"
-    },
-    "API_KEY": {
-      "type": "string",
-      "description": "API key to use. Must be generated in the Bank Frick UI using the PRIVATE_KEY below."
-    },
-    "PRIVATE_KEY": {
-      "type": "string",
-      "description": "RSA key used to produce and verify signatures when authorizing the client. Can be a raw multiline string or a base64 encoded string."
-    },
-    "PAGE_SIZE": {
-      "type": "number",
-      "description": "The number of accounts to fetch per call to /accounts. Must be >= 1 and <= 500.",
-      "default": 500
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
