--- packages/sources/bank-frick/src/config/includes.json
@@ -1,1 +1,0 @@
-{}
