--- packages/sources/bank-frick/src/config/limits.json
@@ -1,10 +1,0 @@
-{
-  "http": {
-    "production": {
-      "rateLimit1s": 1,
-      "rateLimit1m": 30,
-      "note": "rate limits specified here are best guesses, the docs don't specify them"
-    }
-  },
-  "ws": {}
-}
