--- packages/sources/bank-frick/src/config/symbols.json
@@ -1,1 +1,0 @@
-{}
