+++ packages/sources/bank-frick/src/types.d.ts
