--- packages/sources/bank-frick/test/e2e/README.md
@@ -1,1 +1,0 @@
-Tests that run the adapter and send an end to end request using live API endpoints.
