+++ packages/tsconfig.json
@@ -156,9 +156,6 @@
       "path": "./sources/bank-frick"
     },
     {
-      "path": "./sources/bank-frick-test"
-    },
-    {
       "path": "./sources/bea"
     },
     {
