+++ packages/tsconfig.test.json
@@ -156,9 +156,6 @@
       "path": "./sources/bank-frick/tsconfig.test.json"
     },
     {
-      "path": "./sources/bank-frick-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/bea/tsconfig.test.json"
     },
     {
